pz12/watch_store/requirements.txt
Django>=4.0,<5.0
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
Pillow
parso==0.8.3